address

+++ handle_p.c
@@ -7,6 +7,6 @@ int handle_p(void *p)
 	if (p == NULL)
 		m += write(1, "(nil)", 5);
 	else
-		m += handle((unsigned int)p, 16, 2);
+		m += handle((unsigned long)p, 16, 2);
 	return (m);
 }
